### Pull request checklist

- [x] Did you bump the project version?
- [x] Did you add a description to the Pull Request?
- [x] Did you add new tests?
- [x] Did you add reviewers?

### Old deser
```
@benchmark deser_json(ExchangeInfoData, json_str)
BenchmarkTools.Trial: 7 samples with 1 evaluation.
 Range (min … max):  788.652 ms … 850.092 ms  ┊ GC (min … max): 27.89% … 31.28%
 Time  (median):     820.066 ms               ┊ GC (median):    30.80%
 Time  (mean ± σ):   822.370 ms ±  20.313 ms  ┊ GC (mean ± σ):  30.34% ±  1.43%

  █                  █         █ █            █   █           █  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁█▁█▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁█▁▁▁▁▁▁▁▁▁▁▁█ ▁
  789 ms           Histogram: frequency by time          850 ms <

 Memory estimate: 667.81 MiB, allocs estimate: 4027315.
```

### New deser
```
@benchmark deser_json(ExchangeInfoData, json_str)
BenchmarkTools.Trial: 18 samples with 1 evaluation.
 Range (min … max):  276.343 ms … 308.855 ms  ┊ GC (min … max): 0.00% … 2.33%
 Time  (median):     288.023 ms               ┊ GC (median):    1.47%
 Time  (mean ± σ):   289.823 ms ±   9.065 ms  ┊ GC (mean ± σ):  1.39% ± 1.41%

  ▁    ▁ ▁ ▁    █▁▁  ▁    █     ▁▁         ▁  ▁ ▁▁            ▁  
  █▁▁▁▁█▁█▁█▁▁▁▁███▁▁█▁▁▁▁█▁▁▁▁▁██▁▁▁▁▁▁▁▁▁█▁▁█▁██▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  276 ms           Histogram: frequency by time          309 ms <

 Memory estimate: 31.92 MiB, allocs estimate: 394033.
```

```
julia> versioninfo()
Julia Version 1.11.1
Commit 8f5b7ca12ad (2024-10-16 10:53 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 20 × 12th Gen Intel(R) Core(TM) i7-12700H
  WORD_SIZE: 64
  LLVM: libLLVM-16.0.6 (ORCJIT, alderlake)
Threads: 1 default, 0 interactive, 1 GC (on 20 virtual cores)
```